Small fixes: Renamed test dir. Make structs internal.
